restructure deploy action
It is cleaner to run all tasks in one matrix.
We add `main` as a version but sync to the server without a version subdirectory.
We exclude the version subdirectories from the `main` rsync --delete because matrix jobs are not guaranteed to run in a particular order.
If we don't do this it can wipe out the versions after they are copied over.